api/requirements.txt
neurosity == 2.1.1
Flask
Flask-Cors
Flask-RESTful
gunicorn
requests
python-dotenv
r2r[all]